(feat)separate database setup and table setup
