+++ package.json
@@ -6,13 +6,9 @@
   "scripts": {
     "dev": "webpack -d --watch",
     "build": "webpack -p",
-<<<<<<< HEAD
     "test": "mocha '**/*.spec.js'",
     "start": "webpack -p && node server.js",
     "start:watch": "nodemon server.js"
-=======
-    "test": "echo \"Error: no test specified\" && exit 1"
->>>>>>> (feat) set up database schema
   },
   "repository": {
     "type": "git",
@@ -36,20 +32,12 @@
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
-<<<<<<< HEAD
     "chai": "^3.5.0",
     "mocha": "^3.3.0",
     "supertest": "^3.0.0",
-=======
-    "webpack": "^2.4.1",
-    "babel-core": "^6.24.1"
-  },
-  "dependencies": {
     "bookshelf": "^0.10.3",
     "knex": "^0.12.9",
     "mysql": "^2.13.0",
-    "react": "^15.5.4",
->>>>>>> (feat) set up database schema
     "react-dom": "^15.5.4"
   },
   "engines": {
